Migrate server routes to plugins Make robots.txt dynamic
