+++ package.json
@@ -87,7 +87,6 @@
     "debug": "^3.0.1",
     "enzyme-adapter-react-16": "^1.0.4",
     "fetch-ponyfill": "^4.1.0",
-    "good-squeeze": "^5.0.1",
     "hapi": "^17.0.0",
     "hapi-react-views": "^10.0.0",
     "inert": "^5.0.1",
@@ -165,8 +164,8 @@
     "lsp": "heroku releases -a my-app",
     "lss": "heroku releases -a my-app-staging",
     "devtools-server": "echo; echo; echo 'Hit Cmd+Ctrl+ArrowUp to open remote monitoring in Chrome'; echo; echo; remotedev --hostname=localhost --port=8001",
-    "dev": "npm run yarn && rimraf build && concurrently --kill-others -n 'APP,FLOW' -p '[{name}]' 'cross-env DEBUG=my-app* DEBUG_COLORS=true SHOW_ERRORS=true NODE_ENV=development nodemon -x node server/dev.index.js' 'npm run --silent watch-flow'",
-    "dev-mock": "npm run yarn && rimraf build && concurrently --kill-others -n 'APP,FLOW' -p '[{name}]' 'cross-env USE_MOCK_API=true DEBUG=my-app* DEBUG_COLORS=true SHOW_ERRORS=true NODE_ENV=development nodemon -x node server/dev.index.js' 'npm run --silent watch-flow'",
+    "dev": "npm run yarn && rimraf build && concurrently --kill-others -n 'APP,FLOW' -p '[{name}]' 'cross-env DEBUG=my-app* DEBUG_COLORS=true SHOW_ERRORS=true NODE_ENV=development nodemon -x node server/index.dev.js' 'npm run --silent watch-flow'",
+    "dev-mock": "npm run yarn && rimraf build && concurrently --kill-others -n 'APP,FLOW' -p '[{name}]' 'cross-env USE_MOCK_API=true DEBUG=my-app* DEBUG_COLORS=true SHOW_ERRORS=true NODE_ENV=development nodemon -x node server/index.dev.js' 'npm run --silent watch-flow'",
     "dev-a": "npm run yarn && react-native run-android && npm run devtools-server",
     "dev-i": "npm run yarn && react-native run-ios && npm run devtools-server",
     "flow": "flow; test $? -eq 0 -o $? -eq 2",
