--- static/robots.txt
@@ -1,3 +1,0 @@
-# http://www.robotstxt.org
-User-agent: *
-Disallow: /health
